requirements.txt
datasets>=2.17.0
pandas>=2.2.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
tqdm>=4.66.0
ipykernel>=6.29.0
notebook>=7.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
torch>=2.2.0
transformers>=4.38.0
peft>=0.8.2
accelerate>=0.27.0
evaluate>=0.4.1
huggingface_hub>=0.20.0
safetensors>=0.4.2

# Synthetic Data Generation
distilabel>=0.1.0
tenacity>=8.2.0
openai>=1.12.0

# Dev / Quality Assurance
pre-commit>=3.6.0
ruff>=0.2.1
nbstripout>=0.7.1
